+++ src/main/java/learn/thyme/thyteams/team/TeamId.java
@@ -1,0 +1,18 @@
+package learn.thyme.thyteams.team;
+
+import io.github.wimdeblauwe.jpearl.AbstractEntityId;
+
+import java.util.UUID;
+
+public class TeamId extends AbstractEntityId<UUID> {
+
+   /**
+   * Default constructor for JPA
+   */
+   protected TeamId() {
+   }
+
+   public TeamId(UUID id) {
+       super(id);
+   }
+}