+++ src/main/java/learn/thyme/thyteams/team/TeamRepositoryCustom.java
@@ -1,0 +1,5 @@
+package learn.thyme.thyteams.team;
+
+public interface TeamRepositoryCustom {
+    TeamId nextId();
+}