+++ src/main/java/learn/thyme/thyteams/user/UserService.java
@@ -1,5 +1,6 @@
 package learn.thyme.thyteams.user;
 
+import com.google.common.collect.ImmutableSortedSet;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
@@ -23,4 +24,6 @@ public interface UserService {
     void deleteAllUsers();
 
     long countUsers();
+
+    ImmutableSortedSet<UserNameAndId> getAllUsersNameAndId();
 }